chore: clean up codebase
