+++ examples/README.md
@@ -1,0 +1,3 @@
+# Examples
+
+Few example use cases to try out changes, use tests if you want to properly test things.
