+++ package.json
@@ -5,7 +5,7 @@
 	"author": "Lab Digital",
 	"type": "module",
 	"bin": {
-		"intl-extractor": "./dist/cli.cjs"
+		"intl-extractor": "./dist/bin/cli.js"
 	},
 	"scripts": {
 		"build": "tsup",
@@ -19,6 +19,9 @@
 	"dependencies": {
 		"glob": "^10.4.2",
 		"yargs": "^17.7.2"
+	},
+	"peerDependencies": {
+		"typescript": ">=5.5"
 	},
 	"devDependencies": {
 		"@changesets/cli": "^2.27.6",
