+++ src/main.ts
@@ -1,19 +1,13 @@
 import * as glob from "glob";
 import * as fs from "node:fs";
-import { extractLabelsFromFile } from "./parser";
-
-/**
- * Recursive type for labels which can be nested objects containing strings
- */
-type LabelData = {
-	[key: string]: string | LabelData;
-};
+import { updateLabelCache } from "./cache";
+import { extractLabelsFromFile } from "./extract";
+import { type LabelData } from "./types";
 
 /**
  * Main function that collects labels, source file and writes it to the output
  * @param rootPath Root path of typescript files to check
  * @param output JSON file to use for output labels
- *
  */
 export async function processFiles(
 	directories: Array<string>,
@@ -52,70 +46,10 @@ export async function processFiles(
 		if (Object.keys(data).length > 0) {
 			console.info(`Updating labels for ${file}`);
 			// This might not be performant as we do existign source look ups for every added file
-			updateCache({ cache, data, source });
+			updateLabelCache({ cache, data, source });
 		}
 	}
 
 	// Write the new output
 	fs.promises.writeFile(output, JSON.stringify(cache, null, "\t") + "\n");
-}
-
-/**
- * Update existing cache based on given data and source labels
- */
-export function updateCache({
-	cache,
-	source,
-	data,
-}: {
-	cache: LabelData;
-	source: LabelData;
-	data: Record<string, Set<string>>;
-}) {
-	for (const [key, values] of Object.entries(data)) {
-		// Next-intl uses dot notation for nested objects
-		const keys = key.split(".");
-		let currentCache = cache;
-
-		// Set up the namespace in the cache
-		for (let i = 0; i < keys.length; i++) {
-			const currentKey = keys[i];
-			currentCache[currentKey] = currentCache[currentKey] || {};
-			currentCache = currentCache[currentKey] as LabelData;
-		}
-
-		// Add values for each label, try the existing source first
-		// or use the namespace with name as a value
-		for (const value of values) {
-			currentCache[value] =
-				getLabelFromData(source, [...keys, value]) || `${key}.${value}`;
-		}
-	}
-}
-
-/**
- * Traverse through label data and find existing label or return undefined
- * @param path Array of keys to traverse through
- * @param source Label data object to get label from
- * @returns String value if available or undefined if not
- */
-function getLabelFromData(
-	source: LabelData,
-	path: Array<string>
-): string | undefined {
-	let current: LabelData | string = source;
-	for (const key of path) {
-		if (
-			typeof current !== "object" ||
-			current[key] === undefined ||
-			!(key in current)
-		) {
-			return undefined;
-		}
-
-		current = current[key];
-	}
-
-	// Only return label if it's actually a string
-	return typeof current === "string" ? current : undefined;
 }
