+++ tsconfig.json
@@ -1,23 +1,28 @@
 {
-	"$schema": "https://json.schemastore.org/tsconfig",
-	"display": "Backend",
 	"compilerOptions": {
+		// Enable latest features
+		"lib": ["ESNext", "DOM"],
+		"target": "ESNext",
+		"module": "ESNext",
+		"moduleDetection": "force",
+		"jsx": "react-jsx",
 		"allowJs": true,
-		"composite": true,
-		"esModuleInterop": true,
-		"experimentalDecorators": true,
-		"isolatedModules": true,
-		"module": "ES2022",
+
+		// Bundler mode
 		"moduleResolution": "bundler",
+		"allowImportingTsExtensions": true,
+		"verbatimModuleSyntax": true,
 		"noEmit": true,
-		"outDir": "dist",
-		"preserveWatchOutput": true,
-		"resolveJsonModule": true,
-		"sourceMap": true,
-		"skipLibCheck": true,
+
+		// Best practices
 		"strict": true,
-		"strictPropertyInitialization": false,
-		"target": "ES2022"
+		"skipLibCheck": true,
+		"noFallthroughCasesInSwitch": true,
+
+		// Some stricter flags (disabled by default)
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"noPropertyAccessFromIndexSignature": true
 	},
-	"exclude": ["node_modules/*", "**/node_modules/*"]
+	"include": ["./src/**/*"]
 }
